pyproject.toml
[tool.poetry]
name = "atlaskafkaeventlistener"
version = "0.1.0"
description = ""
authors = ["Pengfei Liu <liu.pengfei@hotmail.fr>"]

[tool.poetry.dependencies]
python = "^3.8"

kafka-python = "^2.0.2"
atlaspyapi = "^0.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
